Continue Atreus revamp
Experimenting with removing the right Shift key, as that frees up the
key to make a symbol layer available. If I can get used to a single
Shift key again, this will allow me to separate the number and symbol
layers completely.

A nice bonus of doing things this way is that there will be no "dead"
keys on the symbol layer any more. Previously, I entered the symbol
layer with the I key (right ring finger due to Colemak). That meant
that my right ring finger was completely unusable in that layer, and
my right pinky was awkward at best. Putting this on a thumb key means
that I regain the usage of those fingers.