Clean up Ergodox layout and add tap-hold mods
This change adds Shift keys back to the inner "thumb" keys again, which
matches the Atreus layout. I still use the pinky Shift keys on a regular
basis on the Ergodox, but every now and then when my Atreus muscle
memory takes over, this makes sure that those Shift keys work as
expected as well.

This also adds tap-hold modifiers to the bottom row of keys again.
I've experimented this before and found it not to my liking, but I'm
going to give it another try now that I've reduced the TAPPING_TERM
setting from where it was before.